requirements.txt
google-cloud-vision
selenium
numpy
opencv-python
requests